Spit it up into smaller components
