+++ src/App.svelte
@@ -7,11 +7,11 @@
 <DarkMode />
 <Content class="full-height">
 	<Grid noGutter>
-		<h1 style="margin-bottom: 2rem;">PR Math</h1>
+		<h1 class="mb-06">PR Math</h1>
 		<Row class="mb-08">
 			<Column lg={8}>
-				<p class="mb-05">What is the chance that your pull request gets accepted?</p>
-				<p>As effective software engineers we want our pull requests to get accepted quickly such that we can ship incremental changes to production and increase the value of our software to our users.</p>
+				<p class="mb-05 bx--body-short-02">What is the chance that your pull request gets accepted?</p>
+				<p>As effective software engineers we want our pull requests to get accepted quickly such that we can continuously deliver incremental value to production.</p>
 			</Column>
 		</Row>
 		<Math />
